Move LimeDevUtility.java into right package
