+++ modules/lime-api/src/main/java/de/sprax2013/lime/LimeDevUtility.java
@@ -1,4 +1,4 @@
-package de.sprax2013.lime.configuration;
+package de.sprax2013.lime;
 
 import java.util.logging.Logger;
