+++ modules/lime-api/src/main/java/de/sprax2013/lime/configuration/Config.java
@@ -1,5 +1,6 @@
 package de.sprax2013.lime.configuration;
 
+import de.sprax2013.lime.LimeDevUtility;
 import de.sprax2013.lime.configuration.validation.EntryValidator;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
